[Bugfix] Changed bool handling to byte
The gcc compiler seems to use byte as the value for bool.
Previous implementation used int which was causing the bool
value not to read reliably.

+++ convert.py
@@ -370,7 +370,7 @@ class FuncDeclVisitor(c_ast.NodeVisitor):
     "long" : "long",
     "char" : "byte",
     "void" : "void",
-    "bool" : "int",
+    "bool" : "byte", #  <-- GCC uses byte for this
     "Bool" : "int",
     "_Bool" : "int",
     "uint" : "int",
